// src/models/DescCen2.ts
export interface DescCen2 {
  IdDescCen2: number;
  nameDescCen2: string | null;
  codeDescCen2: string | null;
}

export interface DescCen2Create {
  nameDescCen2: string | null;
  codeDescCen2: string | null;
}

export interface DescCen2Update {
  nameDescCen2: string | null;
  codeDescCen2: string | null;
}